Issd-Backend/requirements.txt
django 
djangorestframework 
django-environ 
djangorestframework-simplejwt 
psycopg2-binary 
django-cors-headers
psycopg2-binary
django-filter